Extract video data loading into useVideos hook
