+++ frontend/src/pages/Home/Home.jsx
@@ -1,4 +1,5 @@
-import { useState, useEffect, useMemo, useRef } from 'react'
+import { useEffect, useMemo, useRef } from 'react'
+import { useVideos } from '@/hooks/frontend/useVideos'
 import { useParams, useNavigate, useSearchParams } from 'react-router-dom'
 import RequestDownloadDialog from '@/components/frontend/RequestDownloadDialog/RequestDownloadDialog'
 import DeleteVideoDialog from '@/components/frontend/DeleteVideoDialog/DeleteVideoDialog'
@@ -21,8 +22,7 @@ export default function Home() {
     const { id } = useParams()
     const [searchParams, setSearchParams] = useSearchParams()
     const navigate = useNavigate()
-    const [videos, setVideos] = useState([])
-    const [playlists, setPlaylists] = useState([])
+    const { videos, setVideos, playlists } = useVideos()
     const videoRef = useRef(null)
     const requestDownloadDialogRef = useRef(null)
     const deleteVideoDialogRef = useRef(null)
@@ -81,48 +81,6 @@ export default function Home() {
         }
     }
     const currentPlaylist = useMemo(() => playlists.find(p => p.some(v => v.id === selectedVideo?.id)) ?? [], [playlists, selectedVideo])
-
-    useEffect(() => {
-        (async () => {
-            const res = await fetch(`${API_URL}/videos`)
-            const data = (await res.json())
-
-            const partRe = /^(.+) part(\d{2})$/
-
-            const videos = data.map(v => ({
-                ...v,
-                visible: parseInt(v.name.match(partRe)?.[2] ?? '1') <= 1,
-                savedTime: localStorage.getItem(`time_${v.id}`)
-            }))
-
-            const videoGroups = Object.groupBy(videos, v => v.name.match(partRe)?.[1] ?? v.name)
-            for (const [base, items] of Object.entries(videoGroups)) {
-                const primary = items.find(v => v.name.match(partRe)?.[2] === '01') ?? items[0]
-                primary.name = base
-                primary.videoCount = items.length
-            }
-
-            const playlistsArr = Object.values(videoGroups)
-                .filter(items => items.length > 1)
-                .map(items => [...items]
-                    .sort((a, b) => {
-                        const na = parseInt(a.name.match(/part(\d{2})$/)?.[1] ?? '0')
-                        const nb = parseInt(b.name.match(/part(\d{2})$/)?.[1] ?? '0')
-                        return na - nb
-                    })
-                )
-
-            setPlaylists(playlistsArr)
-            setVideos(videos)
-
-            Promise.allSettled(videos.map(async v => {
-                const res = await fetch(`${API_URL}/duration/${v.id}`)
-                if (!res.ok) return
-                const { duration } = await res.json()
-                v.duration = duration
-            })).then(() => setVideos(prev => [...prev]))
-        })()
-    }, [])
 
     useEffect(() => {
         function handleKeyDown(e) {
